@@ -3,21 +3,24 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@material-ui/core": "^4.9.8",
-    "@material-ui/icons": "^4.9.1",
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.3.2",
-    "@testing-library/user-event": "^7.1.2",
-    "axios": "^0.26.1",
-    "particles-bg": "2.5.0",
-    "react": "^16.13.1",
-    "react-countdown": "2.3.2",
-    "react-dom": "^16.13.1",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "^5.0.0",
-    "react-select": "^3.1.0",
-    "react-select-country-list": "^2.2.1",
-    "react-tooltip": "^4.2.8"
+    "@mui/material": "^5.11.0", 
+    "@emotion/react": "^11.11.0",
+    "@emotion/styled": "^11.11.0",
+    "@mui/icons-material": "^5.11.0",
+    "@mui/styles": "^5.11.0",
+    "@testing-library/jest-dom": "^6.6.3",
+    "@testing-library/react": "^16.3.0",
+    "@testing-library/user-event": "^14.6.1",
+    "axios": "^1.9.0",
+    "particles-bg": "2.5.5",
+    "react": "^19.1.0",
+    "react-countdown": "2.3.6",
+    "react-dom": "^19.1.0",
+    "react-router-dom": "^7.6.0",
+    "react-scripts": "^5.0.1",
+    "react-select": "^5.10.1",
+    "react-select-country-list": "^2.2.3",
+    "react-tooltip": "^5.28.1"
   },
   "scripts": {
     "start": "react-scripts start",
